Improved style and renamed files.
